[Preload] Escape "$" characters to fix preload build.
During the ui5-preload gulp task, a String.replace is used
to generate the content of the component-preload file.
Hence, any "$$" sequence is interpreted by the String.replace
function as a single "$" character. In order to avoid this,
we can escape the "$" character, like so: "\$". At the end,
this is the exact same "$" character!

Conflicts:
	MES/WebContent/components/disruptions/Formatter.js